remove shakespeare dependency from Yesod.Core
