+++ yesod-core/Yesod/Core.hs
@@ -21,14 +21,15 @@ module Yesod.Core
     , ErrorResponse (..)
       -- * Utitlities
     , maybeAuthorized
-    , widgetToPageContent
+    -- FIXME: API breakage
+    -- , widgetToPageContent
       -- * Defaults
-    , defaultErrorHandler
+    -- FIXME: API breakage
+    -- , defaultErrorHandler
     , defaultYesodMiddleware
     , authorizationCheck
       -- * Data types
     , AuthResult (..)
-    , unauthorizedI
       -- * Logging
     , defaultMakeLogger
     , defaultMessageLoggerSource
@@ -58,13 +59,11 @@ module Yesod.Core
     , loadClientSession
     , Header(..)
     -- * JS loaders
-    , ScriptLoadPosition (..)
-    , BottomOfHeadAsync
       -- * Subsites
     , MonadHandler (..)
-    , MonadWidget (..)
     , getRouteToParent
-    , defaultLayoutSub
+    -- FIXME
+    -- , defaultLayoutSub
       -- * Misc
     , yesodVersion
     , yesodRender
@@ -77,9 +76,7 @@ module Yesod.Core
     , module Yesod.Core.Content
     , module Yesod.Core.Dispatch
     , module Yesod.Core.Handler
-    , module Yesod.Core.Widget
     , module Yesod.Core.Json
-    , module Text.Shakespeare.I18N
     , module Yesod.Core.Internal.Util
     , module Text.Blaze.Html
     , MonadTrans (..)
@@ -93,31 +90,30 @@ module Yesod.Core
       -- * Utilities
     , showIntegral
     , readIntegral
+    -- FIXME: API breakage
       -- * Shakespeare
       -- ** Hamlet
-    , hamlet
-    , shamlet
-    , xhamlet
-    , HtmlUrl
+    -- , hamlet
+    -- , shamlet
+    -- , xhamlet
+    -- , HtmlUrl
       -- ** Julius
-    , julius
-    , JavascriptUrl
-    , renderJavascriptUrl
+    -- , julius
+    -- , JavascriptUrl
+    -- , renderJavascriptUrl
       -- ** Cassius/Lucius
-    , cassius
-    , lucius
-    , CssUrl
-    , renderCssUrl
+    -- , cassius
+    -- , lucius
+    -- , CssUrl
+    -- , renderCssUrl
     ) where
 
 import Yesod.Core.Content
 import Yesod.Core.Dispatch
 import Yesod.Core.Handler
 import Yesod.Core.Class.Handler
-import Yesod.Core.Widget
 import Yesod.Core.Json
 import Yesod.Core.Types
-import Text.Shakespeare.I18N
 import Yesod.Core.Internal.Util (formatW3 , formatRFC1123 , formatRFC822)
 import Text.Blaze.Html (Html, toHtml, preEscapedToMarkup)
 
@@ -139,10 +135,6 @@ import Control.Monad.Trans.Control (MonadBaseControl (..))
 
 import Control.Monad.Trans.Resource (MonadResource (..))
 import Yesod.Core.Internal.LiteApp
-import Text.Hamlet
-import Text.Cassius
-import Text.Lucius
-import Text.Julius
 import Network.Wai (Application)
 
 runFakeHandler :: (Yesod site, MonadIO m) =>
@@ -153,12 +145,6 @@ runFakeHandler :: (Yesod site, MonadIO m) =>
                -> m (Either ErrorResponse a)
 runFakeHandler = Yesod.Core.Internal.Run.runFakeHandler
 {-# DEPRECATED runFakeHandler "import runFakeHandler from Yesod.Core.Unsafe" #-}
-
--- | Return an 'Unauthorized' value, with the given i18n message.
-unauthorizedI :: (MonadHandler m, RenderMessage (HandlerSite m) msg) => msg -> m AuthResult
-unauthorizedI msg = do
-    mr <- getMessageRender
-    return $ Unauthorized $ mr msg
 
 yesodVersion :: String
 yesodVersion = showVersion Paths_yesod_core.version
@@ -178,10 +164,12 @@ maybeAuthorized r isWrite = do
 getRouteToParent :: Monad m => HandlerT child (HandlerT parent m) (Route child -> Route parent)
 getRouteToParent = HandlerT $ return . handlerToParent
 
+{-
 defaultLayoutSub :: Yesod parent
                  => WidgetT child IO ()
                  -> HandlerT child (HandlerT parent IO) Html
 defaultLayoutSub cwidget = widgetToParentWidget cwidget >>= lift . defaultLayout
+-}
 
 showIntegral :: Integral a => a -> String
 showIntegral x = show (fromIntegral x :: Integer)
